+++ CHANGELOG.md
@@ -12,10 +12,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Fixed** for any bug fixes.
 - **Security** in case of vulnerabilities.
 
-## [Unreleased] - Started: 4-7-20
+## [Unreleased] - Started: 4-22-20
+
+## [2.2.0] - Started: 4-7-20
+
+### Added
+- `Orion`: Initial Payer Historical Lives Core to Dev materializations.
+- `Orion`: Payer project `updatePtps` and `removePtps`
 
 ### Changed
+- `Node Scripts, Polaris`: Seeding of treatment plans, org.tps, etc., now uses access collections as entry, instead of a mix of additional criteria _and_ access.
 - `Orion`: Separate Lives and Access sections of historical uploadSheet resolver.
+- `Orion`: Improve Payer Historical Sheet to Core validation for each sheet type.
+- `Orion`: Payer project query resolver I/O changes -- no longer handling pagination on backend.
 
 ### Removed
 - `Polaris`: Removed deprecated `payerCombinedStateLives` usage across the repo: old import process, new import process, custom data export for regional targeting data